add: move interfaces in single file, create Room and PoomUsers interfaces
