+++ src/utils/sendJsonMessage.ts
@@ -1,4 +1,4 @@
-import { User } from "src/models/user"
+import { User } from "src/models/interfaces"
 
 export const sendJsonMessage = (type: string, data: User) => {
   return JSON.stringify({
